# docs/reference/api.yaml
swagger: "2.0"
info:
  title: Tsuru
  description: Open source, extensible and Docker-based Platform as a Service (PaaS)
  # Warning:
  # DO NOT UPDATE THIS FIELD MANUALLY. RUN "make release version=X.Z.Y" INSTEAD.
  version: "1.30"
schemes:
- http
- https
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
paths:
...
  /1.0/platforms/{platform}:
    parameters:
    - name: platform
      in: path
      required: true
      type: string
      minLength: 1
      description: Platform name.
    delete:
      operationId: PlatformDelete
      description: Delete platform.
      responses:
        "200":
          description: Platform removed
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/ErrorMessage"
        "404":
          description: Not found
          schema:
            $ref: "#/definitions/ErrorMessage"
      tags:
      - platform
      security:
      - Bearer: []
    put:
      operationId: PlatformUpdate
      description: Update platform.
      parameters:
      - name: dockerfile_content
        in: formData
        required: false
        type: file
        description: Dockerfile to build the new version from. When omitted and disabled is not set, the platform is rebuilt from the Dockerfile its latest version was built from.
      - name: disabled
        in: formData
        required: false
        type: string
      produces:
      - application/x-json-stream
      consumes:
      - multipart/form-data
      responses:
        "200":
          description: Platform updated
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/ErrorMessage"
        "404":
          description: Not found
          schema:
            $ref: "#/definitions/ErrorMessage"
      tags:
      - platform
      security:
      - Bearer: []
  /1.0/platforms:
    get:
      operationId: PlatformList
      description: List platforms.
      produces:
      - application/json
      responses:
        "200":
          description: Platform list
          schema:
            type: array
            items:
              $ref: "#/definitions/Platform"
        "204":
          description: No content
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/ErrorMessage"
      tags:
      - platform
      security:
      - Bearer: []
    post:
      operationId: PlatformAdd
      description: Add new platform.
      produces:
      - application/x-json-stream
      consumes:
      - multipart/form-data
      parameters:
      - name: name
        in: formData
        required: true
        type: string
      - name: dockerfile_content
        in: formData
        required: true
        type: file
      responses:
        "200":
          description: Platform created
        "400":
          description: Invalid data
          schema:
            $ref: "#/definitions/ErrorMessage"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/ErrorMessage"
      tags:
      - platform
      security:
      - Bearer: []
  /1.6/platforms/{platform}:
    parameters:
    - name: platform
      in: path
      required: true
      type: string
      minLength: 1
      description: Platform info.
    get:
      operationId: PlatformInfo
      description: Platform info.
      produces:
      - application/json
      responses:
        "200":
          description: Platform info
          schema:
            $ref: "#/definitions/PlatformInfo"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/ErrorMessage"
        "404":
          description: Not found
          schema:
            $ref: "#/definitions/ErrorMessage"
      tags:
      - platform
      security:
      - Bearer: []
  /1.6/platforms/{platform}/rollback:
    parameters:
    - name: platform
      in: path
      required: true
      type: string
      minLength: 1
      description: Platform name.
    post:
      operationId: PlatformRollback
      description: Platform rollback.
      parameters:
      - name: image
        in: query
        required: true
        type: string
      produces:
      - application/x-json-stream
      responses:
        "200":
          description: Ok
        "400":
          description: Invalid data
          schema:
            $ref: "#/definitions/ErrorMessage"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/ErrorMessage"
        "404":
          description: Not found
          schema:
            $ref: "#/definitions/ErrorMessage"
      tags:
      - platform
      security:
      - Bearer: []
...
definitions:
...
  ErrorMessage:
    description: Error message.
    type: string
...
  Platform:
    type: object
    properties:
      name:
        type: string
      disabled:
        type: boolean
      source:
        type: string
        description: Dockerfile the latest version was built from. Only returned to users allowed to create or update platforms.
  PlatformInfo:
    type: object
    properties:
      platform:
        type: object
        $ref: "#/definitions/Platform"
      images:
        type: array
        items:
          type: string
...